Created static Range_Query class for shared range queries between TRACE-Q and the DB API
