+++ src/querying/Range_Query_Test.hpp
@@ -2,35 +2,20 @@
 #define TRACE_Q_RANGE_QUERY_TEST_HPP
 
 #include "Query.hpp"
+#include "Range_Query.hpp"
 
 namespace spatial_queries {
 
     class Range_Query_Test : public Query {
-        struct Window {
-            double x_low{};
-            double x_high{};
-            double y_low{};
-            double y_high{};
-            long double t_low{};
-            long double t_high{};
-        };
-
         /**
          * The window that describes the area of the range query.
          */
-        Window window{};
+        Range_Query::Window window{};
 
         /**
          * Describes whether the original trajectory is in the window.
          */
         bool original_in_window{};
-
-        /**
-         * Determines whether the given trajectory is in the window.
-         * @param trajectory Trajectory to check whether is in the window.
-         * @return A boolean value determining whether the given trajectory is in the window.
-         */
-        [[nodiscard]] bool in_range(data_structures::Trajectory const& trajectory) const;
 
     public:
 
@@ -45,10 +30,9 @@ namespace spatial_queries {
          * @param t_high The high time value of the range query.
          */
         Range_Query_Test(const data_structures::Trajectory& original_trajectory, double x_low, double x_high, double y_low,
-                         double y_high, long double t_low, long double t_high) : window{x_low, x_high, y_low, y_high, t_low, t_high}
-        {
-            original_in_window = in_range(original_trajectory);
-        }
+                         double y_high, long double t_low, long double t_high) :
+                         window{x_low, x_high, y_low, y_high, t_low, t_high},
+                         original_in_window{Range_Query::in_range(original_trajectory, window)} {}
 
         /**
          * Evaluates if the original and simplified trajectory return the same result when performing the range query.
